// x-pack/platform/plugins/shared/observability_ai_assistant/common/types.d.ts
export declare enum MessageRole {
    System = "system",
    Assistant = "assistant",
    User = "user",
    Function = "function",
    Elastic = "elastic"
}
export declare enum KnowledgeBaseEntryRole {
    AssistantSummarization = "assistant_summarization",
    UserEntry = "user_entry",
    Elastic = "elastic"
}
export interface PendingMessage {
    message: Message['message'];
    aborted?: boolean;
    error?: any;
}
export interface Deanonymization {
    start: number;
    end: number;
    entity: {
        class_name: string;
        value: string;
        mask: string;
    };
}
// ...
export interface Message {
    '@timestamp': string;
    message: {
        content?: string;
        deanonymizations?: Deanonymization[];
        name?: string;
        role: MessageRole;
        function_call?: {
            name: string;
            arguments?: string;
            trigger: MessageRole.Assistant | MessageRole.User | MessageRole.Elastic;
        };
        data?: string;
    };
}
export interface Conversation {
    '@timestamp': string;
    user?: {
        id?: string;
        name: string;
    };
    conversation: {
        id: string;
        title: string;
        last_updated: string;
    };
    systemMessage?: string;
    messages: Message[];
    labels: Record<string, string>;
    numeric_labels: Record<string, number>;
    namespace: string;
    public: boolean;
    archived?: boolean;
}
type ConversationRequestBase = Omit<Conversation, 'user' | 'conversation' | 'namespace'> & {
    conversation: {
        title: string;
        id?: string;
    };
};
export type ConversationCreateRequest = ConversationRequestBase;
export type ConversationUpdateRequest = ConversationRequestBase & {
    conversation: {
        id: string;
    };
};
export interface KnowledgeBaseEntry {
    '@timestamp': string;
    id: string;
    title?: string;
    text: string;
    type?: 'user_instruction' | 'contextual';
    public: boolean;
    labels?: Record<string, string>;
    role: KnowledgeBaseEntryRole;
    user?: {
        name: string;
    };
    confidence?: 'low' | 'medium' | 'high';
    is_correction?: boolean;
}
export interface Instruction {
    id: string;
    text: string;
}
export declare enum KnowledgeBaseType {
    UserInstruction = "user_instruction",
    Contextual = "contextual"
}